Now everything works
There were some errors in how the indices of the active vertexes were updated, now this doesn't happen anymore and also the idle indices are updated accordingly. There was a stupid numeric error on how the dz was calculated, even if it was near 0 it gave +-2*pi and so all likelihoods went to 0, fixed with a saturation. Deleted sensed=[0,0] cause it meant errors and after those condition that caused these errors were fixed it had no sense existing. Last things left to do are tricycle model, animation and the WLS thingy.